requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
click==8.1.7
et-xmlfile==1.1.0
gitdb==4.0.11
GitHubKit==0.11.3
GitPython==3.1.41
h11==0.14.0
hishel==0.0.24
httpcore==1.0.5
httpx==0.27.0
idna==3.7
Jinja2==3.1.3
loguru==0.7.2
MarkupSafe==2.1.5
numpy==1.26.4
openpyxl==3.1.2
pandas==2.2.2
pydantic==2.7.0
pydantic_core==2.18.1
python-dateutil==2.9.0.post0
pytz==2024.1
setuptools==69.0.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
typing_extensions==4.11.0
tzdata==2024.1
wheel==0.42.0
